Commit 03 - deployment settings and changes to the service.

+++ src/main/java/br/com/fernanda/wsRESTful_Jersey_Hibernate/bean/IProductBean.java
@@ -10,6 +10,6 @@ public interface IProductBean {
     void update (Product product) throws Exception;
     void remove (Product product) throws Exception;
     Product findByName(String Name) throws Exception;
-    Product findById (long id) throws Exception;
+    Product findById (int id) throws Exception;
     List<Product> findAll () throws Exception;
 }
